requirements.txt
openai
langchain-community
langchain-openai
python-dotenv
streamlit
tavily-python
pydpf
pymupdf
langchain-chroma